Request to update the article
We see that we can check the Audit log for "Forms" in this artile however we have found that we are not able to do so currently as mentioned by the customers.  We confirmed this on our demo tenant as well.

It seems that "Forms" is not mentioned in the list after the sentence "Because you can search for the following types of user and admin activity in Office 365" at the first paragraph, could you please remove "Forms" from the list  after "It can take up to 30 minutes or up to 24 hours after an event occurs for the corresponding audit log entry to be displayed in the search results. The following table shows the time it takes for the different services in Office 365."?

Thanks in advance!